refactor(browser): reorganize browser feature files into map and schematic subdirectories
Move browser-related dialog and image classes into dedicated map and schematic subdirectories for better organization. Update imports and file paths accordingly. Also update documentation to reflect the new structure.
